+++ tests/test_pass.c
@@ -51,6 +51,7 @@ static uint32_t test_valid_pass(void) {
     assert(cfg1.min_len == 8);
     assert(cfg1.max_len == 128);
     assert(cfg1.salt_len == 16);
+    assert(cfg1.version == 1);
     assert(cfg1.algorithm == LIBREAUTH_PASS_ARGON2);
     assert(cfg1.length_calculation == LIBREAUTH_PASS_CHARACTERS);
     assert(cfg1.normalization == LIBREAUTH_PASS_NFKC);
@@ -64,6 +65,7 @@ static uint32_t test_valid_pass(void) {
     assert(cfg2.min_len == 8);
     assert(cfg2.max_len == 128);
     assert(cfg2.salt_len == 16);
+    assert(cfg2.version == 1);
     assert(cfg2.algorithm == LIBREAUTH_PASS_ARGON2);
     assert(cfg2.length_calculation == LIBREAUTH_PASS_CHARACTERS);
     assert(cfg2.normalization == LIBREAUTH_PASS_NFKC);
@@ -88,10 +90,12 @@ static uint32_t test_nist_pass(void) {
     assert(cfg1.min_len == 8);
     assert(cfg1.max_len == 128);
     assert(cfg1.salt_len == 16);
+    assert(cfg1.version == 1);
     assert(cfg1.algorithm == LIBREAUTH_PASS_PBKDF2);
     assert(cfg1.length_calculation == LIBREAUTH_PASS_CHARACTERS);
     assert(cfg1.normalization == LIBREAUTH_PASS_NFKC);
     assert(cfg1.standard == LIBREAUTH_PASS_NIST80063B);
+    cfg1.version = 42;
 
     ret = libreauth_pass_hash(&cfg1, password, storage, LIBREAUTH_PASSWORD_STORAGE_LEN);
     assert(ret == LIBREAUTH_PASS_SUCCESS);
@@ -101,6 +105,7 @@ static uint32_t test_nist_pass(void) {
     assert(cfg2.min_len == 8);
     assert(cfg2.max_len == 128);
     assert(cfg2.salt_len == 16);
+    assert(cfg2.version == 42);
     assert(cfg2.algorithm == LIBREAUTH_PASS_PBKDF2);
     assert(cfg2.length_calculation == LIBREAUTH_PASS_CHARACTERS);
     assert(cfg2.normalization == LIBREAUTH_PASS_NFKC);
@@ -125,6 +130,7 @@ static uint32_t test_invalid_pass(void) {
     assert(cfg.min_len == 8);
     assert(cfg.max_len == 128);
     assert(cfg.salt_len == 4);
+    assert(cfg.version == 1);
     assert(cfg.algorithm == LIBREAUTH_PASS_PBKDF2);
     assert(cfg.length_calculation == LIBREAUTH_PASS_CHARACTERS);
     assert(cfg.normalization == LIBREAUTH_PASS_NFKC);